fix to incorrect output in task2

+++ conversion.c
@@ -5,7 +5,7 @@
  * @n: integer to convert
  * @p: to count size of n in binary
  */
-void convert_b(int n, int *p)
+void convert_b(long int n, int *p)
 {
 	if (n < 0)
 	{
@@ -27,7 +27,7 @@ void convert_b(int n, int *p)
  *
  * Return: size of n in binary
  */
-int binary(int n)
+int binary(long int n)
 {
 	int count = 1;
