+++ main.h
@@ -12,9 +12,9 @@ int print_char(char);
 int print_string(char *);
 int print_integer(int n);
 int _putchar(char c);
-int binary(int n);
+int binary(long int n);
 int check_space(const char *str);
 int _printf_r(const char *str, va_list *my_list);
-void convert_b(int n, int *p);
+void convert_b(long int n, int *p);
 void print(int n, int *p);
 #endif
